Refactor :Use new vue version
