+++ package.json
@@ -1,7 +1,6 @@
 {
-  "name": "appfront",
+  "name": "reserve.utsa",
   "version": "0.0.0",
-  "private": true,
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -9,28 +8,29 @@
     "lint": "eslint . --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "@date-io/date-fns": "^3.0.0",
-    "@date-io/dayjs": "^3.0.0",
-    "@date-io/luxon": "^3.0.0",
-    "@mdi/font": "^7.0.96",
-    "@vueuse/core": "^10.7.2",
-    "axios": "^1.6.5",
-    "core-js": "^3.29.0",
-    "date-fns": "^3.2.0",
-    "dayjs": "^1.11.10",
-    "luxon": "^3.4.4",
+    "@mdi/font": "7.0.96",
+    "@vueuse/core": "^10.11.0",
+    "axios": "^1.7.2",
+    "core-js": "^3.34.0",
     "roboto-fontface": "*",
-    "vue": "^3.2.0",
-    "vue-router": "^4.0.0"
+    "vue": "^3.4.0",
+    "vuetify": "^3.5.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^4.0.0",
-    "eslint": "^8.37.0",
-    "eslint-plugin-vue": "^9.3.0",
-    "sass": "^1.60.0",
-    "unplugin-fonts": "^1.0.3",
-    "vite": "^4.5.3",
-    "vite-plugin-vuetify": "^1.0.2",
-    "vuetify": "^3.4.11"
+    "@vitejs/plugin-vue": "^5.0.4",
+    "eslint": "^8.57.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.22.0",
+    "sass": "^1.71.1",
+    "unplugin-fonts": "^1.1.1",
+    "unplugin-vue-components": "^0.26.0",
+    "unplugin-vue-router": "^0.8.4",
+    "vite": "^5.1.5",
+    "vite-plugin-vuetify": "^2.0.3",
+    "vue-router": "^4.3.0"
   }
 }
