+++ src/App.vue
@@ -1,5 +1,10 @@
 <template>
-  <router-view />
+  <v-app>
+    <AppBar></AppBar>
+    <v-main>
+      <router-view />
+    </v-main>
+  </v-app>
 </template>
 
 <script setup>
