--- src/components/Signin.vue
@@ -1,7 +1,0 @@
-<template>
-  
-</template>
-
-<script setup>
-  //
-</script>
