--- src/layouts/default/Default.vue
@@ -1,12 +1,0 @@
-<template>
-  <v-app>
-    <default-bar />
-
-    <default-view />
-  </v-app>
-</template>
-
-<script setup>
-  import DefaultBar from './AppBar.vue'
-  import DefaultView from './View.vue'
-</script>
