--- src/layouts/default/View.vue
@@ -1,9 +1,0 @@
-<template>
-  <v-main>
-    <router-view />
-  </v-main>
-</template>
-
-<script setup>
-  //
-</script>
