+++ src/pages/verify.vue
