+++ src/plugins/vuetify.js
@@ -11,17 +11,9 @@ import 'vuetify/styles'
 // Composables
 import { createVuetify } from 'vuetify'
 
-
 // https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
 export default createVuetify({
   theme: {
-    themes: {
-      light: {
-        colors: {
-          primary: '#1867C0',
-          secondary: '#5CBBF6',
-        },
-      },
-    },
+    defaultTheme: 'light',
   },
 })
