--- src/views/Form.vue
@@ -1,7 +1,0 @@
-<template>
-  <Formpage />
-</template>
-
-<script setup>
-  import Formpage from '@/components/Formpage.vue'
-</script>
