--- src/views/Home.vue
@@ -1,7 +1,0 @@
-<template>
-  <Homepage />
-</template>
-
-<script setup>
-  import Homepage from '@/components/Homepage.vue'
-</script>
