--- src/views/Timetable.vue
@@ -1,7 +1,0 @@
-<template>
-  <Timetablepage />
-</template>
-
-<script setup>
-  import Timetablepage from '@/components/Timetablepage.vue'
-</script>
