--- src/views/Verify.vue
@@ -1,7 +1,0 @@
-<template>
-  <Verify />
-</template>
-
-<script setup>
-  import Verify from '@/components/Verifypage.vue'
-</script>
